finished backend

+++ backend/src/main/java/com/venuex/backend/service/NotificationService.java
@@ -1,5 +1,6 @@
 package com.venuex.backend.service;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,51 +10,63 @@ import org.springframework.web.server.ResponseStatusException;
 
 import com.venuex.backend.DTO.NotificationDTO;
 import com.venuex.backend.entities.Notification;
+import com.venuex.backend.entities.User;
 import com.venuex.backend.repository.NotificationRepository;
+import com.venuex.backend.repository.UserRepository;
 
 @Service
 public class NotificationService {
     private final NotificationRepository notificationRepository;
+    private final UserRepository userRepository;
 
     @Autowired
-    public NotificationService(NotificationRepository notificationRepository) {
+    public NotificationService(NotificationRepository notificationRepository, UserRepository userRepository) {
         this.notificationRepository = notificationRepository;
+        this.userRepository = userRepository;
     }
 
     public NotificationDTO mapToDTO(Notification notification) {
         NotificationDTO notificationDTO = new NotificationDTO();
         notificationDTO.setId(notification.getId());
-        notificationDTO.setUserId(notification.getUser().getId());
-        notificationDTO.setType(notification.getType());
+        notificationDTO.setUserName(notification.getUser().getFirstName());
         notificationDTO.setMessage(notification.getMessage());
         notificationDTO.setSentAt(notification.getSentAt());
         
         return notificationDTO;
     }
 
-    public Notification getNotificationById(Integer id) {
-        return notificationRepository.findById(id)
-        .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Notification not found"));
-    }
+    public List<NotificationDTO> getUserNotification(String userEmail) {
+        User user = userRepository.findByEmail(userEmail)
+            .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "User not found"));
+        
+            List<Notification> notifications = notificationRepository.findByUserIdOrderBySentAtDesc(user.getId());
+            List<NotificationDTO> returnDTO = new ArrayList<>();
 
-    public List<Notification> getNotificationByUserId(Integer userId) {
-        return notificationRepository.findByUserId(userId);
-    }
-
-    public List<Notification> getAllNotifications() {
-        return notificationRepository.findAll();
-    }
-
-    public Notification createNotification(Notification notification) {
-        if (notificationRepository.existsById(notification.getId())) {
-            throw new ResponseStatusException(HttpStatus.CONFLICT, "Notification already exists");
+        for (Notification notification : notifications) {
+            returnDTO.add(mapToDTO(notification));
         }
+        return returnDTO;
+    }
+
+    public Notification createNotification(User user, String message) {
+        if (user == null)  {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User not found");
+        }
+
+        if (message == null || message.isBlank()) {
+            throw new ResponseStatusException(HttpStatus.NOT_ACCEPTABLE, "Message cannot be empty");
+        }
+        Notification notification = new Notification(user,Notification.EmailType.EMAIL,message);
         return notificationRepository.save(notification);
     }
 
-    public void deleteNotification(Integer id) {
+    public void deleteNotification(Integer id, String userEmail) {
         Notification existingNotification = notificationRepository.findById(id)
             .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Notification not found"));
+
+        if (!userEmail.equals(existingNotification.getUser().getEmail())) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "User not authorized to delete");
+        }
         notificationRepository.delete(existingNotification);
     }
